+++ cov_tree/__init__.py
@@ -1,7 +1,7 @@
 from .version import __version__
 from .core import (
     missed_lines_str,
-    Path, CovNode, CovModule, CovFile,
+    Path, PathLike, CovNode, CovModule, CovFile,
     build_cov_tree,
 )
 from .print import print_tree, cov_color, get_available_tree_sets