+++ cov_tree/core/__init__.py
@@ -1,3 +1,3 @@
 from .tools import missed_lines_str
-from .node import Path, CovNode, CovModule, CovFile
+from .node import Path, PathLike, CovNode, CovModule, CovFile
 from .builder import build_cov_tree